apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
redis==5.1.1
celery==5.4.0
qdrant-client==1.11.3
httpx==0.27.2
anthropic==0.37.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
imagehash==4.3.1
Pillow==10.4.0
transformers==4.44.2
torch==2.4.1
invisible-watermark==0.2.0
opencv-python-headless==4.10.0.84
python-multipart==0.0.9
pytest-env==1.1.3
langgraph==0.2.60
langchain-core==0.3.24
playwright==1.48.0
beautifulsoup4==4.12.3
